feat(assumptions) : reorganize new assumption fact
Introduce assumptions/facts.py
